add templates
