+++ index.html
@@ -9,7 +9,6 @@
 <body>
 
 <main class="central">
-
 </main>
 
 <script src="js/main.js"></script>
