+++ js/createElement.js
@@ -1,0 +1,7 @@
+const createElement = (template) => {
+  const outer = document.createElement(`div`);
+  outer.innerHTML = template;
+  return outer;
+};
+
+export default createElement;
