+++ js/templates/game-1.js
@@ -1,3 +1,5 @@
+import createElement from '../createElement.js';
+
 const html = `  <header class="header">
     <div class="header__back">
       <button class="back">
@@ -63,3 +65,7 @@ const html = `  <header class="header">
       <a href="https://vk.com/htmlacademy" class="social-link  social-link--vk">Вконтакте</a>
     </div>
   </footer`;
+
+const gameOne = createElement(html);
+
+export default gameOne;
