+++ js/templates/stats.js
@@ -1,3 +1,5 @@
+import createElement from '../createElement.js';
+
 const html = `<header class="header">
     <div class="header__back">
       <button class="back">
@@ -116,3 +118,7 @@ const html = `<header class="header">
       <a href="https://vk.com/htmlacademy" class="social-link  social-link--vk">Вконтакте</a>
     </div>
   </footer>`;
+
+
+const stats = createElement(html);
+export default stats;
